Update latex grammar to latest tag and colours
Update the latex grammar to tag v0.4.0.

Following the list of captures provided by
[zed](https://zed.dev/docs/extensions/languages#syntax-highlighting),
overhaul the highlighting to give a decent out of the box experience
with existing themes.

Modelled after the newer nvim-treesitter config, but removing parts
which are not relevant to zed causing problems with colouring:
- `@spell` and `@nospell` related to spell checking
- `#lua-match` directive giving unintended extra matches
  (zed doesn't use lua for scripting)
- `@markup.*` captures which are not in the captures in the list linked
  above

With latex language items not matching up perfectly with the captures in
the list linked above, choices were made to repurpose some common but
irrelevant captures to other purposes specifically to latex.
A benefit to this over settling with the closest fitting capture
description is that the most common captures have clear contrasting
colours.

Here are some of these choices for future reference to keep consistency:

| Capture    | Usage in LaTeX                                          |
|------------|---------------------------------------------------------|
| @function  | generic commands (`\command`)                           |
| @keyword   | special commands (`\begin`, `\end`, `\section`,...)     |
| @type      | environment names                                       |
| @attribute | square bracket content to commands                      |
| @constant  | reference, citation and other labels                    |
| @number    | mathmode content                                        |
| @string    | file paths                                              |
| @title     | part, chapter, section,... titles                       |

**n.b.** `@tag` in html might be the most analoguous to environment
names in latex. But languages don't usually have both tags and functions
so they do not contrast, and latex has no strict `type`s.

Minor other change: the nvim highlighting config has minor errors with
curly braces being incorrectly (and asymmetrically) highlighted for
special commands. This is fixed here.