Adjust auto-closing brackets
There were a few conflicts in the auto-closing brackets. In particular,
if two opening brackets end with the same character, then the second has
no effect. Double/single quotes, and normal brackets "()" are not used
in LaTeX syntax. Quotations in prose should be done like `this' or
``that'' but only autoclosing ` is necessary as tapping it twice gives
``''. There are "paired brackets" with multiple characters that appear
in LaTeX such as \left\{ and \right\} in mathmode, or \[ \] for a block
equation. However `texlab` provides snippets for all of these, so that
can be used instead and avoid the conflicts mentioned above.