Add blank shell script for feasibily test for complicated tasks
Any task for forward searching would be complicated (namely determining
PDF file name). This would be easier if done from a shell script, as
opposed to a string in json. This test is to see if shell scripts
contained in the extension can be accessed at a location that a task
could find.