Final tweaks for PR
Comparing the last modification of the evince_synctex.py file against a
hardcoded value is more appropriate than what the zed extension API
encourages with github releases. This is because we will want the path
to be predictable so that it can be called from a zed task (i.e. not in
a directory named after the release).